Final review

+++ src/modules/activities/activities.service.ts
@@ -7,31 +7,64 @@ import { PrismaService } from 'src/config/prisma/prisma.service';
 export class ActivitiesService {
   constructor(private prisma: PrismaService) {}
 
-  findAll() {
-    return this.prisma.activities.findMany();
+  async findAll() {
+    return await this.prisma.activities.findMany();
   }
 
-  findOne(id: number) {
-    return this.prisma.activities.findUnique({
+  async findOne(id: number) {
+    return await this.prisma.activities.findUnique({
       where: { id },
     });
   }
 
-  create(createActivityDto: CreateActivityDto) {
-    return this.prisma.activities.create({
+  async create(createActivityDto: CreateActivityDto) {
+    return await this.prisma.activities.create({
       data: createActivityDto,
     });
   }
 
-  update(id: number, updateActivityDto: UpdateActivityDto) {
+  async update(id: number, updateActivityDto: UpdateActivityDto) {
+    const packagesAsociated = await this.prisma.activities.findUnique({
+      where: { id },
+      include: {
+        packages: true,
+      },
+    });
+
+    if (!packagesAsociated) {
+      throw new Error('No se encontro la actividad');
+    }
+
+    if (packagesAsociated.packages.length > 0) {
+      throw new Error(
+        'No se puede actualizar una actividad con paquetes asociados',
+      );
+    }
+
     return this.prisma.activities.update({
       where: { id },
       data: updateActivityDto,
     });
   }
 
-  remove(id: number) {
-    return this.prisma.activities.delete({
+  async remove(id: number) {
+    const packagesAsociated = await this.prisma.activities.findUnique({
+      where: { id },
+      include: {
+        packages: true,
+      },
+    });
+
+    if (!packagesAsociated) {
+      throw new Error('No se encontro la actividad');
+    }
+
+    if (packagesAsociated.packages.length > 0) {
+      throw new Error(
+        'No se puede eliminar una actividad con paquetes asociados',
+      );
+    }
+    return await this.prisma.activities.delete({
       where: { id },
     });
   }
