+++ src/modules/category-services/category-services.service.ts
@@ -26,6 +26,25 @@ export class CategoryServicesService {
   }
 
   async update(id: number, updateCategoryServiceDto: UpdateCategoryServiceDto) {
+    const serviceAsociated = await this.prisma.categoryServices.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        services: true,
+      },
+    });
+
+    if (!serviceAsociated) {
+      throw new Error('No se encontro la categoria');
+    }
+
+    if (serviceAsociated.services.length > 0) {
+      throw new Error(
+        'No se puede actualizar una categoria con servicios asociados',
+      );
+    }
+
     return await this.prisma.categoryServices.update({
       where: {
         id,
@@ -35,6 +54,25 @@ export class CategoryServicesService {
   }
 
   async remove(id: number) {
+    const serviceAsociated = await this.prisma.categoryServices.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        services: true,
+      },
+    });
+
+    if (!serviceAsociated) {
+      throw new Error('No se encontro la categoria');
+    }
+
+    if (serviceAsociated.services.length > 0) {
+      throw new Error(
+        'No se puede eliminar una categoria con servicios asociados',
+      );
+    }
+
     return await this.prisma.categoryServices.delete({
       where: {
         id,
